chore: tsify dialog (#24324)
